services/accounting/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
Django==4.2.4
django-environ==0.11.2
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
